Fix APIs parmeters and CallingServerClient tests.

+++ sdk/communication/Azure.Communication.CallingServer/tests/Infrastructure/CallingServerLiveTestBase.cs
@@ -299,7 +299,7 @@ namespace Azure.Communication.CallingServer.Tests
                     try
                     {
                         var response = await callConnection.CancelAllMediaOperationsAsync().ConfigureAwait(false);
-                        Assert.AreEqual(response.Status, HttpStatusCode.OK);
+                        Assert.AreEqual(response.Status, (int)HttpStatusCode.OK);
                     }
                     catch (Exception ex)
                     {
